Factor out interface contract
